requirements.txt
pycryptodome
pyasn1

# for download_keys.py
requests